+++ src/__tests__/array.test.js
@@ -1,6 +1,6 @@
 // @flow strict
 
-import * as Result from '../Result';
+import * as Result from '../lib/Result';
 import { array, nonEmptyArray } from '../array';
 import { guard } from '../guard';
 import { number } from '../number';
