+++ src/__tests__/compose.test.js
@@ -1,7 +1,7 @@
 // @flow strict
 
-import * as Result from '../Result';
-import { annotate } from '../debrief';
+import * as Result from '../lib/Result';
+import { annotate } from '../lib/Annotation';
 import { compose, map } from '../utils';
 import { guard } from '../guard';
 import { number } from '../number';
