--- src/debrief/index.js
@@ -1,7 +1,0 @@
-// @flow strict
-
-export { annotate } from './Annotation';
-export { indent } from './utils';
-export { summarize } from './summarize';
-export { serialize } from './serialize';
-export type { Annotation } from './Annotation';
