+++ src/instanceOf.js
@@ -1,7 +1,7 @@
 // @flow strict
 
-import * as Result from './Result';
-import { annotate } from './debrief';
+import * as Result from './lib/Result';
+import { annotate } from './lib/Annotation';
 import type { Decoder } from './types';
 
 export function instanceOf<T>(klass: Class<T>): Decoder<T> {
