+++ src/lib/Result.js
