+++ src/lib/__tests__/Result.test.js
